Updated move(), draw(), and checkCollisions(). Added checkApple() and commenting on code.
I did most of the move() method and added the snake to the draw() method. In the actionPerformed() method, I added a running if and called the move(). I also added the checkApple method, without any implementation and added implementation to the checkCollisions() method to see if the snake eats itself or goes off screen. Finally, I also added sufficient comments to most of my code to improve readability. A good habit to grow.